Video 1 version - Made player able to shoot projectiles, refined a few things, added use of vecmath

+++ core/src/com/corosus/game/component/PhysicsData.java
@@ -34,7 +34,7 @@ public class PhysicsData extends Component {
 		
 	}
 	
-	public void initPhysics(int entID, float x, float y, byte categoryBits, byte maskBits) {
+	public void initPhysics(int entID, float x, float y, short categoryBits, short maskBits) {
 		World world = Game_AI_TestBed.instance().getLevel().getWorldBox2D();
 		
 		/*if (world.isLocked()) {
